backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[argon2]
python-jose
pydantic-settings
pydantic[email]
